+++ hexagonal/src/main/java/com/example/hexagonal/application/core/usecase/FindCustomerByIdUseCase.java
@@ -1,0 +1,17 @@
+package com.example.hexagonal.application.core.usecase;
+import com.example.hexagonal.application.core.domain.Customer;
+import com.example.hexagonal.application.ports.out.FindCustomerByIdOutputPort;
+
+public class FindCustomerByIdUseCase {
+
+    private final FindCustomerByIdOutputPort findCustomerByIdOutputPort;
+
+    public FindCustomerByIdUseCase(FindCustomerByIdOutputPort findCustomerByIdOutputPort) {
+        this.findCustomerByIdOutputPort = findCustomerByIdOutputPort;
+    }
+
+    public Customer find(String id) {
+        return findCustomerByIdOutputPort.find(id)
+                .orElseThrow(() -> new RuntimeException("Customer not found with id: " + id));
+    }
+}