+++ CHANGELOG.md
@@ -9,3 +9,4 @@
     - Creating Adapter to insert Client.
     - Creating Controller.
 - CRUD - Read
+    - Creating UseCase.